+++ fitsio/src/fitsfile.rs
@@ -1023,33 +1023,9 @@ mod test {
     use images::ImageType;
     use tables::{Column, ColumnDataType, ColumnDescription};
     use errors::{Error, IndexError, Result};
+    use testhelpers::{duplicate_test_file, with_temp_file};
     use std::path::Path;
     use std::{f32, f64};
-
-    /// Function to allow access to a temporary file
-    fn with_temp_file<F>(callback: F)
-    where
-        F: for<'a> Fn(&'a str),
-    {
-        let tdir = tempdir::TempDir::new("fitsio-").unwrap();
-        let tdir_path = tdir.path();
-        let filename = tdir_path.join("test.fits");
-
-        let filename_str = filename.to_str().expect("cannot create string filename");
-        callback(filename_str);
-    }
-
-    /// Function to create a temporary file and copy the example file
-    fn duplicate_test_file<F>(callback: F)
-    where
-        F: for<'a> Fn(&'a str),
-    {
-        use std::fs;
-        with_temp_file(|filename| {
-            fs::copy("../testdata/full_example.fits", &filename).expect("Could not copy test file");
-            callback(filename);
-        });
-    }
 
     #[test]
     fn test_opening_an_existing_file() {
